+++ routing/index.js
@@ -4,6 +4,7 @@ const channelsFromHints = require('./channels_from_hints');
 const getChainFeesChart = require('./get_chain_fees_chart');
 const getFeesChart = require('./get_fees_chart');
 const getFeesPaid = require('./get_fees_paid');
+const getIgnores = require('./get_ignores');
 const getPastForwards = require('./get_past_forwards');
 const giftRoute = require('./gift_route');
 const ignoreFromAvoid = require('./ignore_from_avoid');
@@ -15,6 +16,7 @@ module.exports = {
   getChainFeesChart,
   getFeesChart,
   getFeesPaid,
+  getIgnores,
   getPastForwards,
   giftRoute,
   ignoreFromAvoid,