+++ tags/index.js
@@ -1,0 +1,3 @@
+const getTags = require('./get_tags');
+
+module.exports = {getTags};